% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clonotypelist.R
\docType{data}
\name{cdrs21LT2}
\alias{cdrs21LT2}
\title{List of the winning and unspecifically reactive clonotypes for 21LT2}
\format{
Each level of the list contains the CDR3s_aa entry for the clonotype
that showed the highest reactivity towards the MUT antigen ("winner") as
well as a non-reactive clonotype as well as a clonotype with stronger IFNG
expression in the WT compared to the MUT situation.
}
\usage{
cdrs21LT2
}
\description{
List of the winning and unspecifically reactive clonotypes for 21LT2
}
\keyword{datasets}
