% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\docType{data}
\name{cell_filt}
\alias{cell_filt}
\title{Filtered cells}
\format{
Named list of cell numbers following the filtering steps described in this vignette.
}
\usage{
cell_filt
}
\description{
Filtered cells
}
\keyword{datasets}
