% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sharedClonotypes.R
\docType{data}
\name{clonotype_ids}
\alias{clonotype_ids}
\title{Table of customized clonotype IDs}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 69 rows and 8 columns.
}
\usage{
clonotype_ids
}
\description{
\code{data.table} with our customized clonotype IDs for ease of
visualization and comparison. I.e., the CDR3s amino acid sequences are re-
placed with arbitrary IDs. Note thate these clonotypes are those that are
found in both conditions of patient 21LT2, i.e. MUT and WT.
}
\details{
See the vignette "abbreviations_of_clonotypes" about how the consoli-
dation and clean up of the TRA/TRB sequences was done.
}
\seealso{
\code{shared_clonotypes}
}
\keyword{datasets}
