% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{extract_top_pw}
\alias{extract_top_pw}
\title{Extract the top enriched pathways following fgsea}
\usage{
extract_top_pw(
  fgsea_result = fg_reactome[[i]],
  n_pw = 5,
  direction = c("up", "down", "either", "both"),
  p_adj_threshold = 0.01
)
}
\arguments{
\item{fgsea_result}{object generated by \code{fgsea::fgsea}}

\item{n_pw}{integer indicating the number of top PWs to be returned; default: 5}

\item{direction}{either one of "up","down","either","both"}

\item{p_adj_threshold}{default: 0.01}
}
\value{
name(s) of pathways
}
\description{
Extract the top enriched pathways following fgsea
}
\examples{
\dontrun{
topPWs <- extract_top_pw(fg_reactome[[1]], n_pw = 5, direction = "both", p_adj_threshold = 0.01)
fgsea::plotGseaTable(pwi[topPWs], stats =  rank.l[[5]], fgseaRes = fg_reactome[[5]],  gseaParam = 0.5)
}

}
