% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\docType{data}
\name{gene_filt}
\alias{gene_filt}
\title{Filtered genes}
\format{
Named list of gene numbers following the filtering steps described in this vignette.
}
\usage{
gene_filt
}
\description{
Filtered genes
}
\keyword{datasets}
