% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gseaResults.R
\docType{data}
\name{gns_dropouts}
\alias{gns_dropouts}
\alias{fgsea_pid_degns.c20}
\alias{fgsea_reactome_degns.c20}
\alias{fgsea_kegg_degns.c20}
\title{fGSEA results of the most reactive clonotype comparing Mut vs WT for PID gene sets}
\format{
data.frame with GSEA results, generated by \code{\link{fsgea::fgsea}}.
 Each row corresponds to a tested pathway. The columns are the following:

 \itemize{
 \item pathway – name of the pathway as in 'names(pathway)';
 \item pval – an enrichment p-value;
 \item padj – a BH-adjusted p-value;
 \item ES – enrichment score, same as in Broad GSEA implementation;
 \item NES – enrichment score normalized to mean enrichment of random samples of the same size;
 \item nMoreExtreme' – a number of times a random gene set had a more extreme enrichment score value;
 \item size – size of the pathway after removing genes not present in 'names(stats)'.
 \item leadingEdge – vector with indexes of leading edge genes that drive the enrichment.
 }

 @examples \dontrun{
 gs <- "c2.cp.pid.v7.0.entrez.gmt"
 pw.pid <- fgsea::gmtPathways(gs)
 fgsea_pid_degns.c20  <- fgsea(
    pathways = pw.pid, stats = degns.c20.rnks,
    minSize = 15, maxSize = 500, nperm = 100000)
 }

data.frame with GSEA results, generated by \code{\link{fsgea::fgsea}}.
 Each row corresponds to a tested pathway. The columns are the following:

 \itemize{
 \item pathway – name of the pathway as in 'names(pathway)';
 \item pval – an enrichment p-value;
 \item padj – a BH-adjusted p-value;
 \item ES – enrichment score, same as in Broad GSEA implementation;
 \item NES – enrichment score normalized to mean enrichment of random samples of the same size;
 \item nMoreExtreme' – a number of times a random gene set had a more extreme enrichment score value;
 \item size – size of the pathway after removing genes not present in 'names(stats)'.
 \item leadingEdge – vector with indexes of leading edge genes that drive the enrichment.
 }

 @examples \dontrun{
 library(reactome.db)
 pw.reactome <- fgsea::reactomePathways(genes = degns.c20$ENTREZID)
fgsea_reactome_degns.c20  <- fgsea(pathways = pw_reactome,
    stats = degns.c20.rnks,
    minSize = 15, maxSize = 500, nperm = 100000)
 }

data.frame with GSEA results, generated by \code{\link{fsgea::fgsea}}.
 Each row corresponds to a tested pathway. The columns are the following:

 \itemize{
 \item pathway – name of the pathway as in 'names(pathway)';
 \item pval – an enrichment p-value;
 \item padj – a BH-adjusted p-value;
 \item ES – enrichment score, same as in Broad GSEA implementation;
 \item NES – enrichment score normalized to mean enrichment of random samples of the same size;
 \item nMoreExtreme' – a number of times a random gene set had a more extreme enrichment score value;
 \item size – size of the pathway after removing genes not present in 'names(stats)'.
 \item leadingEdge – vector with indexes of leading edge genes that drive the enrichment.
 }

 @examples \dontrun{
 gs <- "c2.cp.kegg.v7.0.entrez.gmt"
 pw.kegg <- fgsea::gmtPathways(gs)
 fgsea_kegg_degns.c20  <- fgsea(
   pathways = pw.kegg,
   stats = degns.c20.rnks,
   minSize = 15, maxSize = 500, nperm = 100000)
 }
}
\usage{
data(fgsea_pid_degns.c20)

data(fgsea_reactome_degns.c20)

data(fgsea_kegg_degns.c20)
}
\description{
Results of gene set enrichment analyses using the logFC values
for all genes when comparing MUT vs WT exposed cells for clonotype C20, the
most reactive one.
The gene sets that were tested were the PID gene sets from MSigDB (v.7).

Results of gene set enrichment analyses using the logFC values
for all genes when comparing MUT vs WT exposed cells for clonotype c20.
The gene sets that were tested were the REACTOME gene sets.

Results of gene set enrichment analyses using the logFC values
for all genes when comparing MUT vs WT exposed cells for clonotype c20.
The gene sets that were tested were the KEGG gene sets from MSigDB (v.7).
}
\details{
See the vignette about DEG analyses for details.

See  the vignette about DEG detection for details.

See the vignette about DE gene detection for details.
}
\keyword{datasets}
