% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\name{load_21LT2filt}
\alias{load_21LT2filt}
\title{Load the SCE for the full SCE data set including cells without V(D)J info}
\usage{
sce.filt <- load_21LT2filt()
}
\value{
an SCE object that needs to be assigned to an object in the environment
}
\description{
Load the SingleCellExperiment object that holds the SCE that
only filtered for low-quality genes and cells; irrespective of V(D)J
information.
}
\seealso{
\code{\link{load_21LT2filt}}, \code{\link{load_21LT2filt}}
}
