% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\name{load_21LT2merged}
\alias{load_21LT2merged}
\title{Load the SCE for the batch-corrected merged data set}
\usage{
sce.merged <- load_0606T1merged()
}
\value{
an SCE object that needs to be assigned to an object in the environment
}
\description{
Load the SingleCellExperiment object that holds the batch-corrected
reduced dimensionality results (MUT/WT = batch) based on all cells;
irrespective of the V(D)J information.
}
