% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\name{load_21LT2shared}
\alias{load_21LT2shared}
\title{Load the SCE for those cells that have a "shared" clonotype}
\usage{
sce.shared <- load_21LT2filt()
}
\value{
an SCE object that needs to be assigned to an object in the environment
}
\description{
Load the SingleCellExperiment object that holds the cells
The UMAP coordinates were re-calculated on the reduced subset after removal
of suspected doublets (see the vignette about the filtering).
}
